+++ src/GenomeTrackApply.cpp
@@ -98,7 +98,7 @@ SEXP gmapply(SEXP _intervals, SEXP _fn, SEXP _track_exprs, SEXP _enable_gapply_i
 			if (interval_idx != scanner.last_scope_idx() || scanner.isend()) {
 				SEXP rarg = eval_expr;
 				for (unsigned iexpr = 0; iexpr < num_track_exprs; ++iexpr) {
-					size_t num_vals = vals.front().size();
+					uint64_t num_vals = vals.front().size();
 
 					runprotect(rvars[iexpr]);
 					rprotect(rvars[iexpr] = RSaneAllocVector(REALSXP, num_vals));
@@ -263,7 +263,7 @@ SEXP gmapply_multitask(SEXP _intervals, SEXP _fn, SEXP _track_exprs, SEXP _enabl
 				if (interval_idx != scanner.last_scope_idx() || scanner.isend()) {
 					SEXP rarg = eval_expr;
 					for (unsigned iexpr = 0; iexpr < num_track_exprs; ++iexpr) {
-						size_t num_vals = vals.front().size();
+						uint64_t num_vals = vals.front().size();
 
 						runprotect(rvars[iexpr]);
 						rprotect(rvars[iexpr] = RSaneAllocVector(REALSXP, num_vals));
@@ -309,7 +309,7 @@ SEXP gmapply_multitask(SEXP _intervals, SEXP _fn, SEXP _track_exprs, SEXP _enabl
 
 			for (int i = 0; i < get_num_kids(); ++i) {
 				void *ptr = get_kid_res(i);
-				size_t res_size = get_kid_res_size(i);
+				uint64_t res_size = get_kid_res_size(i);
 
 				if (!res_size)
 					continue;
